% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateInputs.R
\name{assert_scalar_numeric}
\alias{assert_scalar_numeric}
\title{Assert that an object is a single, non-missing numeric value}
\usage{
assert_scalar_numeric(x, arg_name, positive = FALSE)
}
\description{
Assert that an object is a single, non-missing numeric value
}
\keyword{internal}
