% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateInputs.R
\name{assert_string}
\alias{assert_string}
\title{Assert that an object is a single, non-missing character string}
\usage{
assert_string(x, arg_name)
}
\description{
Assert that an object is a single, non-missing character string
}
\keyword{internal}
