% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtPlot.R
\name{cmtPlot}
\alias{cmtPlot}
\title{Plot conditional mean trajectories (CMT)}
\usage{
cmtPlot(
  data_plot_all,
  condi_time2event,
  event_type_variable,
  event_type,
  bio_variable,
  time_variable,
  survival_variable,
  interval_time = 1/12,
  id_variable = "id"
)
}
\arguments{
\item{data_plot_all}{A \code{data.frame} that includes the biomarker used for plotting.
It is utilized to plot conditional mean trajectories (CMT).}

\item{condi_time2event}{Conditional event time, indicating that all patients should
have events at this time in the plot}

\item{event_type_variable}{Competing risks variable indicator name. Set to NULL if
there are no competing risks.}

\item{event_type}{A vector containing the names of all event types.}

\item{bio_variable}{Name of the biomarker variable used for plotting.}

\item{time_variable}{The name of time variable in linear mixed model.}

\item{survival_variable}{Name of the time-to-event outcomes variable.}

\item{interval_time}{The time interval between two time points. Time points are
plotted within the baseline to event time.}

\item{id_variable}{Name of the patient ID column in \code{data_plot_all}. Default is \code{"id"}.}
}
\value{
Conditional mean trajectories plot.
}
\description{
This function generates the Conditional Mean Trajectories (CMT) plot.
All patients in this plot experience events at the same time point,
specified by \code{condi_time2event}. Several evenly spaced time points between
the baseline and \code{condi_time2event} are selected for plotting. Each point is
calculated using the mean value of all patients' biomarker values at that time point.
The interval between two time points is defined by \code{interval_time}
}
\examples{

# example without competing risks

data(pbc3)

pbc.cmt <- cmtPlot(data_plot_all = pbc3, condi_time2event = 5, 
   event_type_variable = NULL, event_type = NULL,
   bio_variable = "serBilir", time_variable = "year", 
   survival_variable = "years", 
   interval_time = 1/12
)

pbc.cmt


# example with competing risks

data(pbc3)

data_plot_all = pbc3[!is.na(pbc3$status4),]

pbc.cmt.cr <- cmtPlot(data_plot_all, condi_time2event = 5, 
   event_type_variable = 'status4', event_type = c("0", "1"),
   bio_variable = "albumin", time_variable = "year", 
   survival_variable = "years", 
   interval_time = 1/4
)

pbc.cmt.cr

}
