% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalYTBio.R
\name{conditionalYTBio}
\alias{conditionalYTBio}
\title{conditional distribution of Y|T, if no competing risk;}
\usage{
conditionalYTBio(
  Y_all,
  time_new,
  bio_i,
  data_predict_all,
  long_fit_all,
  l_i,
  survival_variable,
  time_variable,
  survival_variable_all,
  survival_trans_function
)
}
\arguments{
\item{time_new}{Prediction time add horizon}

\item{bio_i}{Biomarker used to do prediction}

\item{data_predict_all}{This involves a collection of \code{data.frame} objects for
dynamic prediction, each corresponding to a distinct longitudinal outcome.
These data frames should contain the variables specified in \code{long_sub_fixed}
and \code{long_sub_random}. Utilizing a list structure
allows for the incorporation of multiple longitudinal outcomes,
each potentially following different measurement protocols.
In instances where all longitudinal outcomes are recorded at identical
time points across patients, a singular \code{data.frame} object may
be used in a \code{list}. It is presumed that each data frame is
structured in a long format.}

\item{long_fit_all}{Outputs from the model fitting process using the \code{nlme} package,
encompassing the results and parameters obtained from the analysis.}

\item{l_i}{A vector of time points to calculate the conditional probability.}

\item{survival_variable}{Time-to-event outcomes variable name.}

\item{time_variable}{The name of time variable in linear mixed model.}

\item{survival_variable_all}{The name of the transformed time-to-event outcomes variable.}

\item{survival_trans_function}{The transformation function used for time-to-event outcomes,
in the order of \code{survival_variable_all}.}
}
\value{
The output is a list containing probability matrices of \strong{log} densities
(so that they neither overflow nor underflow; see \code{predictRisk()} for how they
are exponentiated). In the presence of
competing risks, this list includes two elements; otherwise,
it contains only one element. Each element within the list is a probability matrix,
with the number of rows (l_i) corresponding to specific time points and
columns representing different patients. Every matrix element represents
the conditional probability derived from the conditional distribution
of longitudinal variable Y given the survival time T without competing risk D
for a particular patient at a specific time point.
}
\description{
This function computes the conditional probability density function of
longitudinal variable Y, given the survival time T without competing risk D.
}
\keyword{internal}
