% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicPrediction.R
\name{dynamicPrediction}
\alias{dynamicPrediction}
\title{Dynamic prediction function}
\usage{
dynamicPrediction(
  data.predict.all,
  long_fit_all,
  survival_fit_all,
  prediction.time,
  horizon,
  time_variable,
  survivalVariableAll,
  survivalTransFunction,
  bandcount1 = 10,
  bandcount2 = 40
)
}
\arguments{
\item{data.predict.all}{This involves a collection of \code{data.frame} objects for
dynamic prediction, each corresponding to a distinct longitudinal outcome.
These data frames should contain the variables specified in \code{LongSubFixed}
and \code{LongSubRandom}. Utilizing a list structure
allows for the incorporation of multiple longitudinal outcomes,
each potentially following different measurement protocols.
In instances where all longitudinal outcomes are recorded at identical
time points across patients, a singular \code{data.frame} object may
be used in a \code{list}. It is presumed that each data frame is
structured in a long format.}

\item{long_fit_all}{Outputs from the model fitting process using the \code{nlme} package,
encompassing the results and parameters obtained from the analysis.}

\item{survival_fit_all}{Results and parameters generated from the model fitting
procedure, utilizing the \code{coxph} function. These outputs include the comprehensive
findings and variables derived from the analysis.}

\item{prediction.time}{Time used to make the prediction.}

\item{horizon}{Prediction horizon.}

\item{time_variable}{The name of time variable in linear mixed model.}

\item{survivalVariableAll}{The name of the transformed time-to-event outcomes variable.}

\item{survivalTransFunction}{The transformation function used for time-to-event outcomes,
in the order of \code{survivalVariableAll}.}

\item{bandcount1}{The number of points used to perform the numerical integral,
from the prediction time to the prediction time plus the horizon.}

\item{bandcount2}{The number of points used to perform the numerical integral,
from the prediction time to infinity.}
}
\value{
A probability matrix, where the rows (l_i) correspond to specific time points
and the columns to individual patients. Each element within the matrix signifies
the probability of a future event occurring, as dynamically predicted for each
patient at each time point.
}
\description{
The time values in the prediction data subset must be less than the
specified \code{prediction.time} which is the prediction time. The time points for
longitudinal repeated measurements must not surpass the prediction time.
}
\examples{

\donttest{
data(pbc3)

data.survival.fitting =  pbc3[!duplicated(pbc3$id), ]

formMarginalSurv = Surv(years, status3) ~ age + sex
formConditionalCR = NULL

survival_fit_all = survivalSub(data.survival.fitting, formMarginalSurv, 
                               formConditionalCR)

LongSubFixed = list(
  "long1" = serBilir ~ year + age + sex +  (years) + (years) * year,  
  "long2" = prothrombin ~ year + age + sex + (years) + (years) * year,  
  "long3" = albumin ~ year + age + age * year + sex + (years) + (years) * year,  
  "long4" = alkaline ~ year + age + sex + (years) + (years) * year, 
  "long5" = SGOT ~ year + age + sex + (years) + (years) * year, 
  "long6" = platelets ~ year + age + sex + (years)  + (years) * year)

LongSubRandom =list(
  "long1" =  ~ year| id,   
  "long2" =  ~ year| id,    
  "long3" =  ~ year| id,    
  "long4" =  ~ year| id,    
  "long5" =  ~ year| id,    
  "long6" =  ~ year| id)

survivalVariableAll = list(
  "Tyears1",  "Tyears2", "Tyears3", "Tyears4"
)

survivalTransFunction = list(
  fun1 = function(x){abs(x - 1)}, 
  fun2 = function(x){abs(x - 3)}, 
  fun3 = function(x){abs(x - 5)}, 
  fun4 = function(x){abs(x - 7)}
)

# Complete case analysis
data.fit.all = list()
for(i in seq_len(length(LongSubFixed))){
  data.fit.all[[i]] = pbc3[pbc3$status3 == 1, ]
}

# fitting longitudinal submodel
long_fit_all = longitudinalSub(data.fit.all, LongSubFixed, LongSubRandom)

i_PID = 2
data.raw.predict.1 = pbc3[pbc3$id == i_PID, ]

data.predict.all = list()
for(i in seq_len(length(LongSubFixed))){
  data.predict.all[[i]] = data.raw.predict.1[data.raw.predict.1$year <= 3,]
}

# predict risk probability
risk.prob = dynamicPrediction(data.predict.all, long_fit_all, survival_fit_all, 
                              prediction.time = 3, 
                              horizon = 3, time_variable = "year",
                              survivalVariableAll, survivalTransFunction,
                              bandcount1 = 10, bandcount2 = 10)

}

}
