% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeEventTime.R
\name{fill_event_time}
\alias{fill_event_time}
\title{Write event times (and their transforms) into longitudinal data}
\usage{
fill_event_time(
  data_fit_all,
  id,
  T_values,
  event_time,
  survival_variable_all = NULL,
  survival_trans_function = NULL
)
}
\description{
Write event times (and their transforms) into longitudinal data
}
\keyword{internal}
