% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icphFit.R
\name{icph_loglik_terms}
\alias{icph_loglik_terms}
\title{Per-subject log-likelihood of the interval-censored PH model}
\usage{
icph_loglik_terms(L, R, exact, eta, lambda, cuts)
}
\description{
Per-subject log-likelihood of the interval-censored PH model
}
\keyword{internal}
