% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icphFit.R
\name{icph_lp}
\alias{icph_lp}
\title{Linear predictor of an interval-censored PH fit}
\usage{
icph_lp(fit, newdata)
}
\arguments{
\item{fit}{An \code{icph.BJM} object.}

\item{newdata}{One row per patient.}
}
\description{
Linear predictor of an interval-censored PH fit
}
\keyword{internal}
