% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalSub.R
\name{is_interval_surv}
\alias{is_interval_surv}
\title{Is the outcome of a survival formula interval-censored?}
\usage{
is_interval_surv(form_marginal_surv, data)
}
\description{
Is the outcome of a survival formula interval-censored?
}
\keyword{internal}
