% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancePlot.R
\name{outcome_cause_label}
\alias{outcome_cause_label}
\title{Legend label for event type k in performancePlot()}
\usage{
outcome_cause_label(survival_fit_all, k)
}
\description{
Legend label for event type k in performancePlot()
}
\keyword{internal}
