% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_BJM}
\alias{print_BJM}
\title{Combined print summary for a fitted BJM}
\usage{
print_BJM(long_fit_all, survival_fit_all, digits = 4)
}
\arguments{
\item{long_fit_all}{Output from \code{\link{longitudinalSub}}.}

\item{survival_fit_all}{Output from \code{\link{survivalSub}}.}

\item{digits}{Number of significant digits. Default is 4.}
}
\value{
Invisibly returns a named list with both fit objects.
}
\description{
Combined print summary for a fitted BJM
}
