% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeEventTime.R
\name{subject_intervals}
\alias{subject_intervals}
\title{Interval endpoints per subject}
\usage{
subject_intervals(data_survival, form_marginal_surv, id, event_type = NULL)
}
\description{
Interval endpoints per subject
}
\keyword{internal}
