% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalSubVar.R
\name{longitudinalSubVar}
\alias{longitudinalSubVar}
\title{Variance-covariance matrix
Reference: package "lmm" and package "joineRML" function \code{mvlme}.}
\usage{
longitudinalSubVar(thetaLong, l, tol.em, verbose)
}
\description{
Variance-covariance matrix
Reference: package "lmm" and package "joineRML" function \code{mvlme}.
}
\keyword{internal}
