% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariance.R
\name{process_variance}
\alias{process_variance}
\title{Construct variance}
\usage{
process_variance(
  num_i,
  time_new,
  bio_i,
  data.predict.all,
  long_fit_all,
  time_variable
)
}
\description{
Construct variance
}
\keyword{internal}
