% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskPlot.R
\name{riskPlot}
\alias{riskPlot}
\title{Plot of risk using dynamic prediction}
\usage{
riskPlot(
  data_predict_all_pre,
  long_fit_all,
  survival_fit_all,
  prediction_time = NULL,
  bio_i = NULL,
  horizon,
  time_variable,
  survival_variable_all,
  survival_trans_function,
  bandcount1 = 10,
  bandcount2 = 10
)
}
\arguments{
\item{data_predict_all_pre}{This involves a collection of \code{data.frame} objects for}

\item{long_fit_all}{Outputs from the model fitting process using the \code{nlme} package,
encompassing the results and parameters obtained from the analysis.}

\item{survival_fit_all}{Results and parameters generated from the model fitting
procedure, utilizing the \code{coxph} function. These outputs include the comprehensive
findings and variables derived from the analysis.}

\item{prediction_time}{Time used to make the prediction.}

\item{bio_i}{Biomarker used to do prediction.}

\item{horizon}{Prediction horizon.}

\item{time_variable}{The name of time variable in linear mixed model.}

\item{survival_variable_all}{The name of the transformed time-to-event outcomes variable.}

\item{survival_trans_function}{The transformation function used for time-to-event outcomes,
in the order of \code{survival_variable_all}.}

\item{bandcount1}{The number of points used to perform the numerical integral,
from the prediction time to the prediction time plus the horizon.}

\item{bandcount2}{The number of points used to perform the numerical integral,
from the prediction time to infinity.}
}
\value{
Plot of risk using dynamic prediction.
}
\description{
This function gives the risk prediction plot.
}
