% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalSub.R
\name{survivalSubInterval}
\alias{survivalSubInterval}
\title{Interval-censored branch of \code{survivalSub()}}
\usage{
survivalSubInterval(
  data_survival_fitting,
  form_marginal_surv,
  form_conditional_cr,
  event_time,
  baseline,
  df
)
}
\description{
Interval-censored branch of \code{survivalSub()}
}
\keyword{internal}
